pyproject.toml
[project]
name = "tana-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "msal>=1.34.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.36.1",
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.19.0",
    "jinja2>=3.1.6",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "pre-commit>=3.5.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
